requirements.txt
asgiref==3.6.0
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
daphne==4.0.0
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.51.0
django-channels==0.7.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
drf-yasg==1.21.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==22.0
pathspec==0.10.3
platformdirs==2.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
zope.interface==5.5.2
